Adds support for optional inspection.
It's been suggested that a solution to the overhead of reflection
(imposed by the import of `inspect`) can be mitigated by moving to a
provider pattern.

0.7.0 allows for the passing of the `inspect` parameter (defaults to
`False`) to the `Slog` class to switch between a provider which outputs
the file name and line number, or a null provider, which outputs a
space.

The API is otherwise backwards-compatible with 0.6.3.