Refactors inspection methods to reduce overhead.
One unnecessary call to the expensive `currentframe` function is
removed, with one remaining. Also removed are many of the redundant
`f_back` call stack navigation invocations. This should reduce overhead
somewhat in code which either has inspection enabled or makes many calls
to `crit` (although if that's the case, overhead's probably the least of
your worries).